Refactor url path parameter parsing to remove chi dependency
For some reason it can't reliably parse uuids out of the path unless the call to chi's chi.URLParam(r, "uid") is explicity wrapped by uuid's validation method i.e. uuid.MustParse(chi.URLParam(r, "uid")). It's easier to just pick off the last part of the path manually and not deal with this moving forward.

I also added tests for the GetJobsHandler which resurfaced this bug again and led to the refactor.